feat: Add total_students endpoint
This commit adds a new endpoint `/total_students` to the `studentsRoutes.js` file and a corresponding controller function `total_students` in the `studentsController.js` file. The endpoint fetches the total number of students from the database using the `Student.count()` method and returns it in the response.

This change enhances the application by providing the ability to retrieve the total number of students, enabling better tracking and analysis of student data.